ecourseapis/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-ckeditor==6.7.3
django-js-asset==3.1.2
djangorestframework==3.16.1
pillow==12.0.0
PyMySQL==1.1.2
sqlparse==0.5.3
tzdata==2025.2
